LabExamples/Lab07_Ses_ve_Muzik/pyproject.toml
[project]
name = "lab07-ses-ve-muzik"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["pygame-ce>=2.5.7"]
